third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-compile-and-run.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>
var test = function()
{
    function printExceptionDetails(exceptionDetails)
    {
        InspectorTest.addResult("exceptionDetails:");
        var text = exceptionDetails.text;
        if (exceptionDetails.exception)
            text += " " + exceptionDetails.exception.description;
        InspectorTest.addResult("   " + text);
        InspectorTest.addResult("   line: " + exceptionDetails.lineNumber + ", column: " + exceptionDetails.columnNumber);

        var stack = exceptionDetails.stackTrace ? exceptionDetails.stackTrace.callFrames : null;
        if (!stack) {
            InspectorTest.addResult("   no stack trace attached to exceptionDetails");
        } else {
            InspectorTest.addResult("   exceptionDetails stack trace:");
            for (var i = 0; i < stack.length && i < 100; ++i) {
                InspectorTest.addResult("       url: " + stack[i].url);
                InspectorTest.addResult("       function: " + stack[i].functionName);
                InspectorTest.addResult("       line: " + stack[i].lineNumber);
            }
        }
    }

    var contextId = UI.context.flavor(SDK.ExecutionContext).id;
    InspectorTest.runDebuggerTestSuite([
        async function testSuccessfulCompileAndRun(next)
        {
            var expression = "var a = 1; var b = 2; a + b; ";
            InspectorTest.addResult("Compiling script");
            var response = await InspectorTest.RuntimeAgent.invoke_compileScript({expression, sourceURL: "test.js", persistScript: true, executionContextId: contextId});

            InspectorTest.assertTrue(!response[Protocol.Error]);
            InspectorTest.assertTrue(!response.exceptionDetails);
            InspectorTest.assertTrue(!!response.scriptId);

            InspectorTest.addResult("Running script");
            response = await InspectorTest.RuntimeAgent.invoke_runScript({scriptId: response.scriptId, executionContextId: contextId, objectGroup: "console", silent: false});
            InspectorTest.assertTrue(!response[Protocol.Error]);
            InspectorTest.assertTrue(!response.exceptionDetails);
            InspectorTest.addResult("Script result: " + response.result.value);
            next();
        },

        async function testRunError(next)
        {
            var expression = "var a = 1; a + c; ";
            InspectorTest.addResult("Compiling script");
            var response = await InspectorTest.RuntimeAgent.invoke_compileScript({expression, sourceURL: "test.js", persistScript: true, executionContextId: contextId});
            InspectorTest.assertTrue(!response[Protocol.Error]);
            InspectorTest.assertTrue(!response.exceptionDetails);
            InspectorTest.assertTrue(!!response.scriptId);

            InspectorTest.addResult("Running script");
            response = await InspectorTest.RuntimeAgent.invoke_runScript({scriptId: response.scriptId, executionContextId: contextId, objectGroup: "console", silent: false});
            InspectorTest.assertTrue(!response[Protocol.Error]);
            InspectorTest.assertTrue(!!response.exceptionDetails);
            printExceptionDetails(response.exceptionDetails);
            next();
        },

        async function testCompileError(next)
        {
            var expression = "}";
            InspectorTest.addResult("Compiling script");
            var response = await InspectorTest.RuntimeAgent.invoke_compileScript({expression, sourceURL: "test.js", persistScript: true, executionContextId: contextId});
            InspectorTest.assertTrue(!response[Protocol.Error]);
            InspectorTest.assertTrue(!!response.exceptionDetails);
            InspectorTest.assertTrue(!response.scriptId);
            printExceptionDetails(response.exceptionDetails);
            next();
        }
    ]);
}
</script>
</head>
<body onload="runTest()">
<p>Tests separate compilation and run.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=89646">Bug 89646.</a>
</body>
</html>
